Put reference stub in stub.rs
