+++ src/main.rs
@@ -9,7 +9,6 @@ use clashlib::outputstyle::OutputStyle;
 use clashlib::stub::StubConfig;
 use clashlib::{solution, stub};
 use directories::ProjectDirs;
-use indoc::indoc;
 use rand::seq::IteratorRandom;
 
 #[derive(Clone)]
@@ -502,37 +501,7 @@ impl App {
             }
             Some(fname) => std::fs::read_to_string(fname)?,
             None if args.get_flag("from-reference") => {
-                const REFERENCE_STUB: &str = indoc! {r##"
-                    read anInt:int
-                    read aFloat:float
-                    read Long:long
-                    read aWord:word(1)
-                    read boolean:bool
-                    read ABC1ABc1aBC1AbC1abc1:int
-                    read STRING:string(256)
-                    read anInt2:int aFloat2:float Long2:long aWord2:word(1) boolean2:bool
-                    loop anInt read x:int
-                    loop anInt read x:int f:float
-                    loop anInt loop anInt read x:int y:int
-                    loopline anInt x:int
-                    loopline anInt w:word(50)
-                    loopline anInt x:int f:float w:word(50)
-                    write result
-
-                    OUTPUT
-                    An output comment
-
-                    write join(anInt, aFloat, Long, boolean)
-
-                    write join(aWord, "literal", STRING)
-
-                    STATEMENT
-                    This is the statement
-
-                    INPUT
-                    anInt: An input comment over anInt
-                "##};
-                REFERENCE_STUB.to_owned()
+                stub::SIMPLE_REFERENCE_STUB.to_owned()
             }
             None => {
                 let handle = self.current_handle()?;
